refactor(extension): Add useRegistry hook
