+++ packages/extension/src/ui/hooks/useDatabase.test.ts
@@ -3,7 +3,6 @@ import { describe, it, expect, vi, beforeEach } from 'vitest';
 
 import { useDatabase } from './useDatabase.ts';
 import { usePanelContext } from '../context/PanelContext.tsx';
-import { parseObjectRegistry } from '../services/db-parser.ts';
 
 vi.mock('../context/PanelContext.tsx', () => ({
   usePanelContext: vi.fn(),
@@ -20,14 +19,12 @@ vi.mock('../services/db-parser.ts', () => ({
 describe('useDatabase', () => {
   const mockCallKernelMethod = vi.fn();
   const mockLogMessage = vi.fn();
-  const mockSetObjectRegistry = vi.fn();
 
   beforeEach(() => {
     vi.clearAllMocks();
     vi.mocked(usePanelContext).mockReturnValue({
       callKernelMethod: mockCallKernelMethod,
       logMessage: mockLogMessage,
-      setObjectRegistry: mockSetObjectRegistry,
     } as unknown as ReturnType<typeof usePanelContext>);
   });
 
@@ -37,7 +34,6 @@ describe('useDatabase', () => {
       expect(result.current).toStrictEqual({
         fetchTables: expect.any(Function),
         fetchTableData: expect.any(Function),
-        fetchObjectRegistry: expect.any(Function),
         executeQuery: expect.any(Function),
       });
     });
@@ -140,66 +136,6 @@ describe('useDatabase', () => {
       await expect(
         result.current.executeQuery('SELECT * FROM test'),
       ).rejects.toThrow('Invalid query');
-    });
-  });
-
-  describe('fetchObjectRegistry', () => {
-    it('should query the kv table and parse the result', async () => {
-      const { result } = renderHook(() => useDatabase());
-      const mockKvData = [
-        { key: 'obj1', value: '{"id":"obj1","type":"test"}' },
-        { key: 'obj2', value: '{"id":"obj2","type":"test"}' },
-      ];
-      const mockParsedData = {
-        gcActions: '',
-        reapQueue: '',
-        terminatedVats: '',
-        vats: {},
-      };
-
-      mockCallKernelMethod.mockResolvedValueOnce(mockKvData);
-      vi.mocked(parseObjectRegistry).mockReturnValueOnce(mockParsedData);
-
-      result.current.fetchObjectRegistry();
-
-      await waitFor(() => {
-        expect(mockCallKernelMethod).toHaveBeenCalledWith({
-          method: 'executeDBQuery',
-          params: { sql: 'SELECT key, value FROM kv' },
-        });
-        expect(parseObjectRegistry).toHaveBeenCalledWith(mockKvData);
-        expect(mockSetObjectRegistry).toHaveBeenCalledWith(mockParsedData);
-      });
-    });
-
-    it('should log errors when fetching object registry fails', async () => {
-      const { result } = renderHook(() => useDatabase());
-      const errorResponse = { error: 'Table not found' };
-      mockCallKernelMethod.mockResolvedValueOnce(errorResponse);
-
-      result.current.fetchObjectRegistry();
-
-      await waitFor(() => {
-        expect(mockLogMessage).toHaveBeenCalledWith(
-          'Failed to fetch object registry: "Table not found"',
-          'error',
-        );
-      });
-    });
-
-    it('should handle promise rejection when fetching object registry', async () => {
-      const { result } = renderHook(() => useDatabase());
-      const error = new Error('Query failed');
-      mockCallKernelMethod.mockRejectedValueOnce(error);
-
-      result.current.fetchObjectRegistry();
-
-      await waitFor(() => {
-        expect(mockLogMessage).toHaveBeenCalledWith(
-          'Failed to fetch object registry: Query failed',
-          'error',
-        );
-      });
     });
   });
 });
